Print all tests line on error for typescript
Previously sanna could only print all tests lines for javascript. It
Would cut of at the wrong point when running in typescript with source
maps on since it was using the length of the transpiled js.

No sanna finds the next line in the stack trace that is from the failing
test file and uses that to find the end of the test.